leitianyu:update(9.5)

+++ src/main/java/com/metabubble/BWC/service/impl/TeamServiceImpl.java
@@ -47,7 +47,7 @@ public class TeamServiceImpl extends ServiceImpl<TeamMapper, Team>
         String phone = MobileUtils.blurPhone(tel);
         //查询是否为会员
         if (userService.checkGrade(id)) {
-            BigDecimal bigDecimalForFirstWithVip = BigDecimal.valueOf(Integer.parseInt(configService.getOnlyContentById(Long.parseLong("18"))));
+            BigDecimal bigDecimalForFirstWithVip = BigDecimal.valueOf(Double.parseDouble(configService.getOnlyContentById(Long.parseLong("18"))));
             if (!bigDecimalForFirstWithVip.equals(0)) {
                 //会员返现
                 team.setTotalWithdrawnAmount(team.getTotalWithdrawnAmount().add(bigDecimalForFirstWithVip));
@@ -56,7 +56,7 @@ public class TeamServiceImpl extends ServiceImpl<TeamMapper, Team>
                 userMsgService.addCashback(id,tel,"一级成员会员返现"+bigDecimalForFirstWithVip);
             }
         }else {
-            BigDecimal bigDecimalForFirstWithNtoVip = BigDecimal.valueOf(Integer.parseInt(configService.getOnlyContentById(Long.parseLong("19"))));
+            BigDecimal bigDecimalForFirstWithNtoVip = BigDecimal.valueOf(Double.parseDouble(configService.getOnlyContentById(Long.parseLong("19"))));
             if (!bigDecimalForFirstWithNtoVip.equals(0)) {
                 //非会员返现
                 BigDecimal add = team.getTotalWithdrawnAmount().add(bigDecimalForFirstWithNtoVip);
@@ -83,7 +83,7 @@ public class TeamServiceImpl extends ServiceImpl<TeamMapper, Team>
 
         //查询是否为会员
         if (userService.checkGrade(id)) {
-            BigDecimal bigDecimalForSecondWithVip = BigDecimal.valueOf(Integer.parseInt(configService.getOnlyContentById(Long.parseLong("20"))));
+            BigDecimal bigDecimalForSecondWithVip = BigDecimal.valueOf(Double.parseDouble(configService.getOnlyContentById(Long.parseLong("20"))));
             if (!bigDecimalForSecondWithVip.equals(0)) {
                 //会员返现
                 team.setTotalWithdrawnAmount(team.getTotalWithdrawnAmount().add(bigDecimalForSecondWithVip));
@@ -92,7 +92,7 @@ public class TeamServiceImpl extends ServiceImpl<TeamMapper, Team>
                 userMsgService.addCashback(id,tel,"二级成员会员返现"+bigDecimalForSecondWithVip);
             }
         }else {
-            BigDecimal bigDecimalForSecondWithNtoVip = BigDecimal.valueOf(Integer.parseInt(configService.getOnlyContentById(Long.parseLong("21"))));
+            BigDecimal bigDecimalForSecondWithNtoVip = BigDecimal.valueOf(Double.parseDouble(configService.getOnlyContentById(Long.parseLong("21"))));
             if (!bigDecimalForSecondWithNtoVip.equals(0)) {
                 //非会员返现
                 team.setTotalWithdrawnAmount(team.getTotalWithdrawnAmount().add(bigDecimalForSecondWithNtoVip));
